Move eval dashboard, bug report, and agent manual into chatbot.
These are product/UI modules, not shared framework. Appearance stays
in framework because scripting/LibrePy import it for Monaco theme sync.

No framework shims: callers, tests, LibrePy bundle paths, and docs
now import plugin.chatbot.* directly.
