+++ docs/repo-map.md
@@ -28,7 +28,8 @@ Start here by task.
 | UNO document helpers | Writer chat-context assembler and `DocumentService` (WriterAgent). Thin dispatcher: Writer via `text_helpers`, Calc via lazy `plugin.calc.analyzer`, Draw/Impress via `plugin.draw.bridge.get_draw_context_for_chat`. | [`plugin/doc/document_helpers.py`](../plugin/doc/document_helpers.py) |
 | Light document helpers | Linebreaks, tracked-deletion reads, heading tree, path, selection text, Writer text slices; type guards; UD props (LibrePy-safe) | [`plugin/doc/text_helpers.py`](../plugin/doc/text_helpers.py), [`plugin/doc/doc_type.py`](../plugin/doc/doc_type.py), [`plugin/doc/udprops.py`](../plugin/doc/udprops.py) |
 | Config / keys / LRU | `writeragent.json`, keys, LRU | [`plugin/framework/config.py`](../plugin/framework/config.py) (I/O, cache, getters); [`plugin/framework/config_schema.py`](../plugin/framework/config_schema.py) (pure schema/coercion — import from here, not `config`) |
-| Dialogs / XDL | Dialog load helpers and settings UI | [`plugin/chatbot/dialogs.py`](../plugin/chatbot/dialogs.py), [`plugin/chatbot/dialog_views.py`](../plugin/chatbot/dialog_views.py), [`plugin/chatbot/settings_dialog.py`](../plugin/chatbot/settings_dialog.py) |
+| Dialogs / XDL | Dialog load helpers and settings UI | [`plugin/chatbot/dialogs.py`](../plugin/chatbot/dialogs.py), [`plugin/chatbot/dialog_views.py`](../plugin/chatbot/dialog_views.py), [`plugin/chatbot/settings_dialog.py`](../plugin/chatbot/settings_dialog.py), [`plugin/chatbot/eval_dashboard_ui.py`](../plugin/chatbot/eval_dashboard_ui.py), [`plugin/chatbot/bug_report.py`](../plugin/chatbot/bug_report.py) |
+| Agent manual | On-demand `get_guidance` topics + agent-backend full manual | [`plugin/chatbot/agent_manual.py`](../plugin/chatbot/agent_manual.py) |
 | Async UI drain | Stream queue drain on the UI thread (`get_toolkit`, `get_ctx`) | [`plugin/framework/async_stream.py`](../plugin/framework/async_stream.py), [`plugin/framework/uno_context.py`](../plugin/framework/uno_context.py) |
 | Writer HTML / apply | HTML import and apply-content paths (callers `import format as format_support`) | [`plugin/writer/format.py`](../plugin/writer/format.py) |
 | Writer charts / shapes | Shared tool names with Calc/Draw; declare union of `uno_services` | [`plugin/writer/specialized/charts.py`](../plugin/writer/specialized/charts.py), [`plugin/writer/specialized/shapes.py`](../plugin/writer/specialized/shapes.py) |
