+++ plugin/chatbot/agent_manual.py
