+++ tests/framework/test_constants.py
@@ -20,7 +20,7 @@ from plugin.framework.prompts import (
     DEFAULT_DRAW_GREETING,
 )
 
-# NOTE: the EXTERNAL_AGENT_GUIDANCE pin test moved to tests/framework/test_agent_manual.py —
+# NOTE: the EXTERNAL_AGENT_GUIDANCE pin test moved to tests/chatbot/test_agent_manual.py —
 # the blob was retired; the single source is the shared prompt pieces in constants.py (the
 # sidebar embeds them, get_guidance serves them per topic, full_manual() feeds the agent backend).
